requirements.txt
together
streamlit
python-dotenv
os